Fix User

+++ Wunderlist/Epam.Wunderlist.Models/User.cs
@@ -5,7 +5,7 @@ namespace Epam.Wunderlist.Models
 {
     public class User 
     {       
-        public Int16 Id { get;protected set; }
+        public int Id { get;protected set; }
         public string UserName { get; set; }
         public string Email { get; set; }
